+++ record_ios/ios/record_ios/Sources/record_ios/Recorder.swift
@@ -75,6 +75,9 @@ class Recorder {
         completionHandler(path)
         self.updateState(RecordState.stop)
       })
+    } else {
+      completionHandler(nil)
+      updateState(RecordState.stop)
     }
   }