+++ record_ios/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.1.1
+fix: Calling stop never ends when not recording.
+
 ## 1.1.0
 * feat: Add AudioInterruptionMode to `RecordConfig`.
 * feat: Add stream buffer size option.