+++ record_macos/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.1.1
+fix: Calling stop never ends when not recording.
+
 ## 1.1.0
 * feat: Add stream buffer size option.
 * fix: Fix echo cancellation (#523)